# packages/api-contract/openapi-v1.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multicountry Pesticide KG API",
    "version": "1.1"
  },
  "paths": {
...
    "/api/v1/schema": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Schema",
        "operationId": "schema_api_v1_schema_get",
        "parameters": [
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SchemaData_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/overview": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Overview",
        "operationId": "overview_api_v1_stats_overview_get",
        "parameters": [
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_OverviewData_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/countries": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Countries",
        "operationId": "countries_api_v1_stats_countries_get",
        "parameters": [
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Release"
            }
          },
          {
            "name": "X-PestKG-Release",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pestkg-Release"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_CountryData__"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or release selector",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Entity, release, or resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Release lifecycle conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation or configured result-limit failure",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "No validated active release is ready",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountryData": {
        "properties": {
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "jurisdiction_name": {
            "type": "string",
            "title": "Jurisdiction Name"
          },
          "sovereign_country": {
            "type": "string",
            "title": "Sovereign Country"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "official_url": {
            "type": "string",
            "title": "Official Url"
          },
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "source_snapshot_eligible": {
            "type": "boolean",
            "title": "Source Snapshot Eligible"
          },
          "source_rows": {
            "type": "integer",
            "title": "Source Rows"
          },
          "skipped_rows": {
            "type": "integer",
            "title": "Skipped Rows"
          },
          "nodes": {
            "type": "integer",
            "title": "Nodes"
          },
          "edges": {
            "type": "integer",
            "title": "Edges"
          },
          "broken_edges": {
            "type": "integer",
            "title": "Broken Edges"
          },
          "graph_scope": {
            "type": "string",
            "title": "Graph Scope"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "iso3": {
            "type": "string",
            "title": "Iso3"
          },
          "map_id": {
            "type": "string",
            "title": "Map Id"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageRecord"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "jurisdiction",
          "jurisdiction_name",
          "sovereign_country",
          "site_id",
          "official_url",
          "source_file",
          "source_sha256",
          "source_snapshot_eligible",
          "source_rows",
          "skipped_rows",
          "nodes",
          "edges",
          "broken_edges",
          "graph_scope",
          "language",
          "iso3",
          "map_id",
          "coverage"
        ],
        "title": "CountryData"
      },
      "CoverageRecord": {
        "properties": {
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "source_language": {
            "type": "string",
            "title": "Source Language"
          },
          "records": {
            "type": "string",
            "title": "Records"
          },
          "crop_source": {
            "type": "string",
            "title": "Crop Source"
          },
          "target_source": {
            "type": "string",
            "title": "Target Source"
          },
          "active_source": {
            "type": "string",
            "title": "Active Source"
          },
          "formulation_source": {
            "type": "string",
            "title": "Formulation Source"
          },
          "crop_english": {
            "type": "string",
            "title": "Crop English"
          },
          "target_english": {
            "type": "string",
            "title": "Target English"
          },
          "active_english": {
            "type": "string",
            "title": "Active English"
          },
          "formulation_english": {
            "type": "string",
            "title": "Formulation English"
          },
          "crop_english_given_source": {
            "type": "string",
            "title": "Crop English Given Source"
          },
          "target_english_given_source": {
            "type": "string",
            "title": "Target English Given Source"
          },
          "active_english_given_source": {
            "type": "string",
            "title": "Active English Given Source"
          },
          "formulation_english_given_source": {
            "type": "string",
            "title": "Formulation English Given Source"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "jurisdiction",
          "source_language",
          "records",
          "crop_source",
          "target_source",
          "active_source",
          "formulation_source",
          "crop_english",
          "target_english",
          "active_english",
          "formulation_english",
          "crop_english_given_source",
          "target_english_given_source",
          "active_english_given_source",
          "formulation_english_given_source"
        ],
        "title": "CoverageRecord"
      },
...
      "Envelope_OverviewData_": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "$ref": "#/components/schemas/OverviewData"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[OverviewData]"
      },
...
      "Envelope_SchemaData_": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "$ref": "#/components/schemas/SchemaData"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[SchemaData]"
      },
...
      "Envelope_list_CountryData__": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.1"
          },
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CountryData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "data"
        ],
        "title": "Envelope[list[CountryData]]"
      },
...
      "ErrorDescriptor": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDescriptor"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptor"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "OverviewData": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "published_at": {
            "type": "string",
            "title": "Published At"
          },
          "cutoff": {
            "type": "string",
            "title": "Cutoff"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "distribution_status": {
            "type": "string",
            "title": "Distribution Status"
          },
          "known_limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Limitations"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "jurisdictions": {
            "type": "integer",
            "title": "Jurisdictions"
          },
          "source_records": {
            "type": "integer",
            "title": "Source Records"
          },
          "country_nodes": {
            "type": "integer",
            "title": "Country Nodes"
          },
          "country_edges": {
            "type": "integer",
            "title": "Country Edges"
          },
          "shared_nodes": {
            "type": "integer",
            "title": "Shared Nodes"
          },
          "alignment_edges": {
            "type": "integer",
            "title": "Alignment Edges"
          },
          "node_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Node Types"
          },
          "relation_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Relation Types"
          },
          "coverage": {
            "items": {
              "$ref": "#/components/schemas/CoverageRecord"
            },
            "type": "array",
            "title": "Coverage"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "title",
          "version",
          "published_at",
          "cutoff",
          "status",
          "distribution_status",
          "known_limitations",
          "mode",
          "jurisdictions",
          "source_records",
          "country_nodes",
          "country_edges",
          "shared_nodes",
          "alignment_edges",
          "node_types",
          "relation_types",
          "coverage"
        ],
        "title": "OverviewData"
      },
...
      "SchemaData": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "node_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Node Fields"
          },
          "edge_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Edge Fields"
          },
          "node_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Node Types"
          },
          "relation_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Relation Types"
          },
          "federation_predicates": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Federation Predicates"
          },
          "rules": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Rules"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "schema_version",
          "node_fields",
          "edge_fields",
          "node_types",
          "relation_types",
          "federation_predicates",
          "rules"
        ],
        "title": "SchemaData"
      }
    }
  }
}